requirements.txt
pandas==2.2.3
pyodbc==5.2.0
SQLAlchemy==2.0.40
pyyaml==6.0.2
streamlit==1.45.0
protobuf==3.20.3
python-dotenv==1.1.0